added stubs for Dashboard sub-components: Feed, Charts
