+++ client/src/containers/Dashboard/Charts.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Charts = () => {
+  return <div>Charts</div>;
+};
+
+export default Charts;
