+++ client/src/containers/Dashboard/Feed.js
@@ -1,0 +1,8 @@
+import React from 'react'
+
+
+const Feed = () => {
+    return <div>Feed</div>
+}
+
+export default Feed
