+++ client/src/containers/index.js
@@ -2,7 +2,7 @@ export { App } from "./App";
 export { Homepage } from "./Homepage";
 export { SignIn } from "./SignIn";
 export { SignUp } from "./SignUp";
-export { Dashboard } from "./Dashboard";
+export { Dashboard } from "./Dashboard/Dashboard";
 export { Journal } from "./Journal/Journal";
 export { AddActivity } from "./AddActivity/AddActivity";
 export { Blog } from "./layout/Footer/Product"
